WL#14388 Push condition referring fields from an SPJ ancestor table
Patch introduce extensions to the NdbQuery (SPJ) API for specifying
how a parameter required by conditions should be constructed.

NdbQueryOptions::setParameters() may be used to specify the parameter
referred by the the parameterized pushed conditions. It should consist
of an array of 'linkedOperands' which is the NdbQuery way of specifying
Column being referred from ancestor tables.

Code has been added to the NdbQueryBuilder to produce a serialized
parameter discriptor when such a parameter has been specified. That
code is build upon an existing pattern already used to specify
'linked' keys, index-bounds and prune keys. (Which will also result
in a similar reuse of SPJ code)

Usage of these new methods to be introduced in the following patches.

Reviewed By: Arnab Ray <arnab.r.ray@oracle.com>