Bug#32980875: WL#14422: ASSERTION `FALSE' FAILED|SQL/BASIC_ROW_ITERATORS.H
In queries that use both DISTINCT and GROUP BY, the hypergraph
optimizer may attempt to do sort-ahead for the interesting grouping
given by DISTINCT. If the SELECT list contains any aggregates, this
could result in a query plan that attempts to sort on an aggregate
before the value of that aggregate has been calculated.

So far, problems have been seen only with queries that use full-text
search. These queries are not yet supported by the hypergraph
optimizer, but we fix it now to avoid problems when support for
full-text search is enabled.

It is fixed by not attempting to sort-ahead on any orderings that
include aggregates, since sort-ahead always happens before the
aggregates have been calculated.

Change-Id: I46d27a835a2a8d364009dc973ebdae4c01d1d020