Bug #32996884: SPLIT RANGE OPTIMIZER INTO SEVERAL FILES [pre-copy, noclose]
This is a preliminary commit to help git track the upcoming split.
It adds a dummy change to opt_range.cc, and then copies it
to a bunch of files in sql/range_optimizer. This helps git
understand that these files are related (since they are touched
in the same commit), which means that git log --follow and
git blame -C will do better at preserving history.

In this commit, the new files are inert, ie., they are not part
of the build system at all.

If git blame takes you to this commit, consider re-running it with the
-C flag for better tracking.

Change-Id: I6d86915001272e4489358a641831670d8393811f