Bug #32996884: SPLIT RANGE OPTIMIZER INTO SEVERAL FILES
Split opt_range.cc into several files in sql/range_optimizer,
making it easier to refactor and reuse the individual parts
for the hypergraph join optimizer.

The nominal destination file is range_optimizer.cc, but approximately
13,000 lines are moved out into other files; mostly by means of having
two files for each type of range scan (one for the execution part,
one for planning/estimation of it).

This is a pure split, ie., no code is added or changed (beyond one or
two instances of “static”) -- git and Gerrit should mostly show it
as a series of deletion, since all of these files were made a copy of
opt_range.cc in the previous commit. (In some cases, reordering of
the individual functions would make sense, but this has also been
kept to a minimum, to reduce the diff for this commit.) This means
that the resulting code is functionally identical.

If git blame takes you to this commit, consider re-running it with the
-C flag for better tracking.

Change-Id: I4fe805b208224c21b8e49e7426f3d447d87502b1