Bug#33003132 ROUTERTEST_METADATA_CACHE_PLUGIN_CONFIG SEGFAULTS ON FREEBSD13
The test that crashes is using typeid() string to feed the
abi::__cxa_demangle function. The problem is that this function does not like
the  value returned by the typeid() which is
'N11mysqlrouter18option_not_presentE'. It returns -2 error, which our wrapper
does not handle correctly.

This patch is getting rid of the call to the demangle function. We compare the
values returned by the typeid() directly now.

Reviewed By: Jan Kneschke <jan.kneschke@oracle.com>
Reviewed By: Lukasz Kotula <lukasz.kotula@oracle.com>