Bug#32783943: Item_field::fix_outer_field: assertion `cur_query_block
There is a call order mismatch in Query_block::prepare_values():
setup_order() is called after resolve_subquery(). This means that for
a VALUES clause that is a subquery, we may merge the subquery into
the outer query block before calling setup_order(). On merging the
subquery, any Item_field objects belonging to the subquery will have
their Name_resolution_context object's query_block member updated.
When actually trying to resolve these items, data structures are
inconsistent and an error is reported.

Fixed by performing setup_order() earlier. In this case, the column
is not found and the resolving is aborted.

Reviewed by: Chaithra Gopala Reddy <chaithra.gopalareddy@oracle.com>

Change-Id: I173495505ee467d015f0114bfe2da68d188be883