Bug #32798288 addendum2 Approved by: Bharathy Satish <bharathy.x.satish@oracle.com>
On windows there's no permissions check hence the my_stat() call
is a no-op and the flags passed to my_open() kick in and print the
error for !include.
But on unixes my_stat is called with 0 for flags and produces no
error. Hence the error for !include is printed on windows and not
printed on unix.
Fixed by passing down the flags to my_stat and replacing the
resulting error message in the test to look like it's been generated
by my_open.