Bug #33003335: HYPERGRAPH: ASSERTION `FD_SET == IT_AND_INSERTED.FIRST->SECOND.ACTIVE_FUNCTIONAL
The hypergraph cycle construction code could inadvertently create a
cycle that involved a STRAIGHT_JOIN, which was unintentional.
This was caught by a test fairly roundabout: When the cycle was
created, we would not correctly promote its join edge to a predicate
(which would give silently wrong results) since we always treat
outer and straight joins as bridges in the graph, and this would in
turn be caught by the interesting order subsystem, which detected
inconsistently applied functional dependencies. (One could discuss
whether the test for bridges is actually needed when promoting
join predicates, but at least, it made this bug show up.)

Fix by checking the companion set of all conditions involved in the
join, not just the one we are making a cycle out of. This should
prevent the cycle in question from being made in the first place.

Change-Id: I6fe28dad89326c521b9720391f83433099d7af30