Bug#33005571 - WL#14533 SEGMENTATION FAULT IN RPDRQC_FIND_QKRN_NODE() AT RPDRQCQK_QKRN.C Bug#33007424 - WL#14533 ASSERTION 'COUNTVISIBLEFIELDS(*TOP_LEVEL_ITEMS) == ROWSET.NCOLS' FAILED Bug#33007730 - WL#14533 ASSERTION FAILURE IN RPDMPR_CAST_AS_VARCHAR_RPDMPR_DTYPE_SB8_RPDMPR Bug#33022692 - WL#14533 SEGMENTATION FAULT IN RPD::EXTRACTDISTINCTNODES AT HA_RPD_QKRN.CC Bug#33034687 - WL#14533 SEGFAULT IN RPD::GETUSEDTABLES AT HA_RPD_QKRN.CC Bug#33041923 - WL#14533 RAPID RESULTSET DIFFERENCE WITH SUM(DATETIME) AND ROLLUP Bug#33041851 - WL#14533 RESULTSET DIFFERENCE WITH QUERY HAVING GROUPING(STR_TO_DATE()) Bug#33050370 - WL#14533 SEGMENTATION FAULT IN RPDMPR_UNARY_MATH_RPDMPR_DTYPE_DOUBLE_RPDMPR
Bug#33005571:
=============
There was an inifinite recursion, because rpdrqc_find_qkrn_node() is
tracking the qkrn node ID which is CONSUMING the shared result. However,
with the ROLLUP transformation, there may be multiple table scans
consuming the same shared result.

The fix is to track the PRODUCER's node ID of a sharing intermediate
result, instead of the CONSUMER's node ID.

Bug#33007424:
=============
MySQL did not call unwrap_rollup_group() on the item in function
find_subquery_in_select_list(). By that, the query block could contain
too many Items which would lead to a crash in RAPID because the field
count between a query block's visble items and the JOIN's visible items
did not match.
Two MTR test cases use
  --replace_result SQ1_field1 field4
because the headline is not yet correctly changed, a separate patch will
be merged which will also remove this the --replace_result intermediate
fix.

Bug#33007730:
=============
The output encoding for GroupBy keys in the original GroupBy's result
vector was not properly set to NULLable.
Fixing this, however, led to QCOMP mixing up partitioning information,
because the GroupBy key qkropn and the respective result vector qkropn
had different encoding::not_null information. To fix that, in
rpdrqcqk_is_qkropn_enc_same() the comparison of
  qkropn_enc::not_null_qkropn_enc
was removed. Adapted the function comment and added a comment which
points to Enh#33048610 for investigating whether this has any other bad
impact --- note that QA did not find any issues related to this change,
though.

Bug#33022692:
=============
Crash was not reproducible, but query did not offload. Reason was that
in PatchRvOpnsForFinalTs(), the check whether we run out of indexes
while patching the final table scan's result vector, was too early.
This must only be checked before the index list iterator is actually
accessed, because it may happen that the index list is exhaustet (at the
end), but some operands are still inspected for which the index list is
not needed. This is not an error case. Moved the check only to where it
is actually needed.
Furthermore, the caller did not check the return bool value and in
case the functions fails, the query would still be offloaded.

Bug#33034687:
=============
Was not reproducible anymore.

Bug#33041923:
=============
The offload check for SUM or AVG over temporal data types was not
kicking in, because the SUM / AVG aggregation item was wrapped inside an
Item_rollup_sum_switcher. Indeed, only very few places where we handled
SUM_FUNC_ITEMs did check whether the aggregation item was a sum switcher
or not. To fix this bug, added function UnwrapSumSwitcher(), which works
like unwrap_rollup_group() and unpacks the argument in case it is a sum
switcher.

Bug#33041851:
=============
For the final ROLLUP branch's FTAG we used a CO(UNT) operator to count
the input rows. However, when there was e.g. only a NULL input row, then
CO(UNT) returns 0, while CO(UNT)ST(A)R also counts NULL values. Due to
that, the final ROLLUP branch did not return any row, while when using
COSTR it _does_ return the desired row.

Bug#33050370:
=============
NULLability information was not always propagated to
Item_rollup_sum_switcher. Fixed by setting nullable based on master
child.

Futher problems fixed / improvements:

* The ROLLUP transformation must not happen on derived tables, because
  then the node IDs may be matching different nodes when transforming
  the outer query blocks. Therefore, add a parameter to the postfix
  transformers which passes down the information whether transformation
  is on an outer query block.
  For that, added function IsOuterQueryBlock() and replaced all checks
  against Query_block::outer_query_block() with that function, as often
  the bare pointer value is used in boolean contexts.

* In ha_rpd_qkrn_expr.cc when translating ROPTTYP_ROLLUP_GROUP and
  ROPTTYP_GROUPING then ConstructQkrnExprTree() can return false while
  the constructed qrkopn is actuall nullptr. This was not considered,
  so added checks for both function types to avoid nullptr access.

* Fix a few MTR tests when running MTR on pb2 with 2 rpdservers:
  + All the .*dataplacement.* tests: source
    use_secondary_engine_forced.inc
    and
    use_secondary_engine_reset.inc
    instead of setting this manually.
  + cp_oom_commit_hooks: make unsorted result sorted.
  + pfs: Do not run with multiple rpd servers. This test is run on
    testdb (not baumeister) only anyways, but with 2 rpdservers it
    produces different results, which is expected.

* To allow for a better offload trace message, added a Heap parameter to
  function ExtractItemFromTmpField() to also print the original
  expression when that function fails.

* Due to one of the bug fixes, more MTR queries got offloaded in
  tests i_rollup and select_{dict|varlen}. As one consequence, merged
  i_rollup.test into rollup.inc and cp_rollup.inc, because that query
  was only offloaded for VARLEN encoding.

* Added many SELECT DISTINCT test cases.

Change-Id: I804128e477348162dec9aa343bbeb22032f66bc2