WL#14673: Enable the hypergraph optimizer for DELETE [patch 1/2, multi-table]
This patch enables use of the hypergraph join optimizer for optimizing
multi-table DELETE statements, when the hypergraph optimizer has been
enabled.

Some extra care had to be taken to preserve row IDs through hash
joins. The old optimizer does not use hash joins for DELETE, whereas
the hypergraph optimizer may use them. After a hash join, the row ID
of tables on the right side of the join can no longer be fetched from
the handler of the underlying table or index scan, because the hash
join returns rows from those tables in a different order than that of
the underlying scan. Row IDs for such tables are therefore stored in
the hash join buffer, and the deletion logic needs to read them from
the buffer instead of the handler. If the hash join spills to disk, we
have the same issue with tables on the left side of the join. Since we
don't know up front if a join will spill to disk, we also store row
IDs for target tables on the left side of a hash join.

Also, the hypergraph optimizer takes into account the cost of creating
temporary tables for "buffered" deletes, and is therefore more likely
to pick a join order where "immediate" deletes can be performed. (That
is, it tries to construct join orders where it is safe to delete rows
immediately when they are returned from the join, instead of
collecting row IDs in temporary tables and postponing the deletion
until all rows have been read from the join.)

Otherwise, once the plan has been generated by the hypergraph
optimizer, and transformed into an iterator tree, the execution
follows the same execution logic as multi-table DELETE statements
optimized by the old optimizer.

Change-Id: Iaf93af9db2b87912c6787cd7d00f6cea225bc51b