Bug #33025668: REMOVE INIT_ALLOC_ROOT()
Remove the legacy init_alloc_root() wrapper. This commit basically
inlines it everywhere, removing the unused prealloc_size parameter
in the process.

Note that we cannot use regular placement new to initialize MEM_ROOTs,
since that would hit the overload to allocate _from_ the MEM_ROOT.
(init_alloc_root() did this, but was put over the overloads, so that
they were not in scope for that specific call.) Thus, in the remaining
few places where we can't just use the regular constructor, we cast
to void * to get the usual placement new.

Change-Id: I331da71d410036468910362b62986c111d4db4d0