WL#14697: Reject calls to MATCH with rollup column as argument
When the input to the MATCH function is a grouping column with rollup,
and the MATCH function needs to be evaluated after aggregation (it's
located in the SELECT list, the GROUP BY clause, the HAVING clause or
the ORDER BY clause), an error will be raised.

The rejected queries used to return wrong results. The fundamental
issue is that MATCH is not implemented as a function of its arguments,
but rather as a function of the current record id. This means that the
usual approach of switching the argument from the base column to a
temporary table column with the rollup value does not work, because
MATCH can only return correct results for base columns.

The change does not affect calls to MATCH from the WHERE clause, since
it only needs to evaluate the function before the aggregation, and
therefore accesses only the base columns.

Change-Id: Ia22f47aa5c59a8bfcdb07f94472acfbeae4f8f35