Bug #33037007: PRELIMINARY FIXES FOR WL #14488 [get_quick_keys friend, noclose]
QUICK_RANGE_SELECT had a strange two-phase construction, where the
object would first be constructed, and then get_quick_keys() would
go and poke into the object's private members to set up the ranges
some other parameters. Run get_quick_keys() first to get the ranges,
then send them to the constructor. This removes the need for it
to be a friend, and simplifies error handling.

Change-Id: I82d7e27e5b3819ba9e7c64d305f21b3235003c3c