Bug#33036184: GROUP BY WITH ROLLUP AND CONCAT OVER GROUP BY               KEY DOES NOT PRODUCE NULL
If a query has rollup, while setting an expression as nullable
because it has a grouping column, we currently miss setting
all the expressions within that expression as nullable. We set
it only for the top-most expression. This leaves some of the
expressions as not nullable (if they were originally deemed
not nullable). As a result, during evaluation, NULL generated
by rollup is not propagated as one of the expression is not
nullable.

For E.g:
If we have f(g(x)), and x is a grouping column, we currently
mark x and f() as nullable and not g(), which is a problem,
since f() may assume that if g() is not nullable, then it
cannot have produced a NULL.

Solution:
We now set all the expressions containing a grouping column as
nullable if we have a query with rollup.

Change-Id: I6675ca8141e663785f21c8afc716e59959ae0533