Bug#33041720: REDUNDANT CALL TO INIT_FTFUNCS() IN MULTI-TABLE DELETE
Multi-table delete statements that use full-text search, call
init_ftfuncs() twice. First in JOIN::optimize_fts_query(), and later
in Query_result_delete::optimize(). This is harmless, but unnecessary.

This patch removes the second call. It also adds a basic test case
which exercises multi-table delete with full-text search, as this
combination was not tested before in the test suite.

Change-Id: I61fb8513c8d591e6568d76e0fdebe077fb307cdd