Bug #18684036: ELIMINATE LAST FEW HEAP USAGE FROM THE RANGE OPTIMIZER
Currently, the range optimizer allocates memory in a myriad of places:

 1. TRPs, SEL_TREEs, a few other things: On its own, internal MEM_ROOT
    (which cannot be reused across calls, even though
    DynamicRangeIterator would like to do so).
 2. A few items and lists: On the per-statement MEM_ROOT; sometimes
    through thd->mem_root (stored in param->old_root), sometimes through
    *THR_MALLOC.
 3. Data used by QUICK_SELECT_I: In one tiny MEM_ROOT owned by the
    QUICK_SELECT_I itself, except when a quick owns another quick,
    in which case the parent's MEM_ROOT is usually used -- except
    when that's disallowed by way of assertion.
 4. QUICK_SELECT_I itself: On the heap, using operator new.

Note that thd->mem_root points to various versions of this throughout
the range optimizer execution.

This is confusing, causes unneccessary heap allocations (especially for
DynamicRangeIterator), and most importantly: It currently makes it
impossible to return TRPs, since they are destroyed before return.

We replace this with a simpler scheme, involving only two MEM_ROOTs:

 1. All temporary objects (SEL_TREE etc.) are allocated on
    temp_mem_root.
 2. Everything that's supposed to be permanent and used after the call
    (TRPs, QUICK_SELECT_I, associated data) is allocated on return_mem_root.
    This is typically equivalent to thd->mem_root.

The caller supplies both, so that DynamicRangeIterator can point both
to the same place and reuse it for each execution. thd->mem_root is
unchanged (so the few Items we may be creating etc. land automatically
in the right place). There's still one heap allocation, of the MRR
buffer; it is large (256 kB), so it does not gain much from being moved
into the MEM_ROOT.

The DynamicRangeIterator now runs the range optimizer while holding
the plan lock, as it clears out the MEM_ROOT (with the old quick)
before invoking it, to be able to reuse the memory.

Note that this moves TRPs from temporary to permanent storage, which is
in line with the plan to use them as planning objects instead of using
QUICK_SELECT_I for that purpose.

The THD argument to make_quick() was in practice used only to change
thd->mem_root, and was thus removed as obsolete. (The THD is still
used once, in index merge, to initialize a Unique object. It was not
deemed worth it versus just using current_thd.)

Change-Id: Icd7b3cea60982cd971dbf308782343c371491798