Bug#32981242 - ACCESSING REST API USING CREDENTIAL PROVIDED BY FILE IS SLOW
Description
===========
Routers REST API is slow when it uses acount informations provided
by `passwd` file. The computation done while credential verification
make the process slow.

Fix
===
To reduce the slowness of computation, the patch is going to
do the computation only once per account. If verification was
successful then its going to cache a hash of the users credentials.
At next authentication, the account is going to be verified against
cached hash, which is much faster.

RB: 26487
Reviewed by: Jan Kneschke <jan.kneschke@oracle.com>