Bug #33037007: PRELIMINARY FIXES FOR WL #14488 [quick_optim, noclose]
Remove the obsolete QEP_TAB::quick_optim; it was only used at a time
where filesort rewrote quick, which we haven't done for a while.

keyread_optim is seemingly still used in EXPLAIN, although it's not
really clear whether we could just remove it and replace it with
table->key_read.

Change-Id: Ie0d84bee6eb88ff2a1cf4094ef3f2fea8753e0bc