fix CREATE...SELECT row-based binlog in ANSI mode
# What is changed
Ignore the ANSI sql mode when generating queries for the binlog.

# Why
Since MySQL 8.0.21, CREATE...SELECT statement are replicated using a single
transaction instead of two. To implement that correctly, the binlog contain a
CREATE TABLE...START TRANSACTION Query followed by Table_map/Write_rows events
to insert the rows found by the SELECT :

    'binlog.000001', 1336, 'Gtid', 1, 1415, "SET @@SESSION.GTID_NEXT= 'c8ec3a3b-ce8d-11eb-a313-8c8caaa4a076:6'"
    'binlog.000001', 1415, 'Query', 1, 1492, 'BEGIN'
    'binlog.000001', 1492, 'Query', 1, 1636, 'use `db_one`; CREATE TABLE "t1" (\n  "1" int NOT NULL DEFAULT \'0\'\n) START TRANSACTION'
    'binlog.000001', 1636, 'Table_map', 1, 1686, 'table_id: 111 (db_one.t1)'
    'binlog.000001', 1686, 'Write_rows', 1, 1726, 'table_id: 111 flags: STMT_END_F'
    'binlog.000001', 1726, 'Xid', 1, 1757, 'COMMIT /* xid=24 */'

This non-standard SQL syntax is correctly recognised as "transactional DDL" and
does not generate an implicit commit in the middle of a binlog transaction.

To write the modified query inside the binlog, the same code used to implement
SHOW TABLES is reused.

In addition to its documented effects, the ANSI sql mode also alters the
behavior of SHOW TABLES to hide MySQL specific syntax (this mode is internally
named "foreign_db_mode"). This is useful to create standard SQL dumps.

Because of the two previous points, when ANSI mode is enabled, the generated
query inside the binlog does not contain the necessary START TRANSACTION.

Without the START TRANSACTION, the replica trying to execute the binlog will
see a normal CREATE TABLE query, which will try to do an implicit commit
in the middle of a binlog transaction and fail with this error :

    Error 'Cannot execute statements with implicit commit inside a transaction when @@SESSION.GTID_NEXT == 'UUID:NUMBER' '.' on query.

To keep the query short, the binlog example at the beginning shows a query
which requires tables without primary keys to be allowed, but the problem is
the same when creating a table with a primary key, or when selecting rows from
an actual table instead of just SELECT 1.