WL#14673: Enable the hypergraph optimizer for DELETE [patch 2/2, single-table]
This patch enables use of the hypergraph join optimizer for optimizing
single-table DELETE statements, when the hypergraph optimizer has been
enabled.

Contrary to the old optimizer, the hypergraph optimizer does not have
a separate code path for optimizing single-table DELETE. It uses the
full capabilities of the hypergraph optimizer to generate the plan.

This also means that it will follow the multi-table DELETE code path
during execution, and never go through
Sql_cmd_delete::delete_from_single_table(). Because of this, it does
not use ha_delete_all_rows() to delete all rows in one go for
statements known to delete all rows in a table with no delete
triggers, and instead always deletes the rows one by one. However,
this method of deleting all rows was not used on InnoDB tables with
the old optimizer either. It was mainly an optimization for MyISAM.

Since single-table DELETE statements may have ORDER BY and LIMIT
clauses, which the multi-table DELETE syntax does not accept, some
adjustments had to be made to the planning of DELETE statements. In
particular, the execution plans needed to preserve row IDs through the
sort node for the ORDER BY clause, so that the correct rows were
deleted.

Change-Id: I7dafde460c36a49840f89dc4d337b51d7cf20361