Bug#33057397: FIELD NAMES ARE DISPLAYED INCORRECTLY               WITH ROLLUP AND SUBQUERY_TO_DERIVED
While transforming a subquery into a derived table,
transformation is expected to replace all instances
of that subquery with a field from the derived table.
However, if the subquery is wrapped in a
rollup_group_item, resolver fails to find the subquery
in the select list which results in adding a hidden
item leading to problems later. An earlier fix ensured
that a rollup group item (Item_rollup_group_item) is
unwrapped before this comparison thereby avoiding the
additional hidden item.

But, if the same subquery is found in group by (if
group by refers to the subquery in the select list
through an alias), even this instance of the subquery
needs to be replaced. However, comparisons fail for
this case and there is a hidden item again.

The second problem is that, the name of the field that
replaces the subquery is not changed to the alias used
for the subquery. Instead it retains the name from the
derived table.

All of these issues arise because we are trying to
compare rollup_group_item(subquery_item) with
(subquery_item) in replace_subquery_in_expr().

Solution:
Before we make a call to replace_subquery_in_expr(),
we unwrap the Item_rollup_group_item and use that for
replacement and comparisons.

Note that, some of the rapid tests need change in their
test files because the names of the columns were from
the derived table instead of the alias in the select
list which is fixed now.

Change-Id: Ic40b183cdeefb574b7cbcb25e628897d29fc1f88