WL#14422: Full-text search in the hypergraph join optimizer
Post-push fix: Xcode 10 doesn't properly support the deduction guides
needed for converting a function pointer to std::function:

join_optimizer.cc:2961:38: error: no viable constructor or deduction
guide for deduction of template arguments of 'function'
                                     std::function(IsFullTextFunction));
                                     ^

Whereas the immediate problem is that the Xcode version used is too
old, it should be unnecessary to wrap the function pointer in
std::function when calling WalkItem. This is currently not possible,
because WalkItem only accepts the function argument to be an rvalue.
If an lvalue is given (for example a reference to a lambda function,
or a raw function pointer), compilation fails:

item.h:2364:33: error: 'type name' declared as a pointer to a
reference of type 'bool (&)(const Item *)'
    return (*reinterpret_cast<T *>(arg))(this);
                                ^

Fixed by making Item::walk_helper_thunk() remove the reference part of
the type so that it doesn't attempt to create a pointer to a reference
type. Similar changes were made to CompileItem/analyzer_helper_thunk
for symmetry.

Now WalkItem() and CompileItem() accept both lvalue references and
rvalue references to function objects or lambdas, as well as raw
function pointers. They do not yet support const function objects or
const lambdas.

Change-Id: Iafe8f3163c6d1a759cfa63468b1bef91df7fd2b7