Bug #33030136: EXPLAIN: ASSERTION `CURRENT_THD->LEX->USING_HYPERGRAPH_OPTIMIZER' FAILED.
In EXPLAIN, if we are crossing into a different query block, but there's
a streaming or materialization node in the way, don't count it as the
root; we want any SELECT printouts to be on the actual root node
(and also the logic for not printing them for ZERO_ROWS to hold).

This does have some unfortunate consequences when a query ends in an
unneeded materialization (the subquery is posted on an unexpected node),
but in general should be positive.

Change-Id: Ic9a5dcd9e711e38d817e8942f282b80e9d4982bb