Bug #33037007: PRELIMINARY FIXES FOR WL #14488 [noclose]
Post-push fix: Don't send in base_min_key and base_max_key as const char *,
as newer GCC assumes that this means they will be read from and most be
initialized. Mark the pointer as const instead to signal to the reader
that they are just a base, in a way that GCC doesn't choke on.

Change-Id: If794649318b53c16b4ee097e46e030705199d0a7