Bug#32731450 MTR SHOULD NOT OVERRIDE PARALLEL SETTING FOR UNIT TESTS
MTR reduces the number of parallel workers internally if there
are fewer tests to run than the value given to --parallel.
Though the reduced value is only applicable for regular tests,
it is currently passed to ctest as well slowing down execution
of unit tests. The patch updates MTR to use the original number
of parallel jobs when running unit tests.

Change-Id: I54f5c4242e7f7edcab74f5eaa372bb914af7b188