Bug#32983024: USE A DIFFERENT HASH FUNCTION FOR GENERATED HIDDEN COLUMN NAMES
The hidden generated columns that back functional indexes, have names
that are generated with the MD5 function. In some hardened
environments, this function is disabled. To allow functional indexes
to be created in these environments too, the naming scheme of the
hidden columns is changed.

Now, the name of the hidden column is on the form:

  !hidden!index_name!key_part_number!counter

The counter part is usually 0, but in the unlikely event that a column
with that name already exists in the table, it is incremented until
the name becomes unique. (Before, adding a functional index would fail
if the generated name was already used by another column.)

Adding a new column with a name used by an existing hidden column, on
the other hand, would fail with a message saying "The column name is
already in use by a hidden column". (Same as before.)

Change-Id: Ib93691db7ea90aab3bb43ef9bd20dcb4cbc7f8e5