Bug#33016004: Problem when calling audit_log_filter_flush() in a prepared stmt
The problem may occur when a UDF is used in a prepared statement, the
UDF is deemed constant, the UDF is also used as argument to some
function and the UDF is evaluated as a constant value inside
the resolve_type() function of that function. What happens is that
during the call to udf_handler::fix_fields(), when prearing a prepared
statement, an early exit is performed to avoid calling the init
function of the UDF. However, in that early exit, a call guard was
not cleared, meaning that the UDF's buffers and handler were freed.

The fix is to call udf_fun_guard.defer(), to avoid freeing the
UDF's buffers and handler, thus ensuring the following evaluation
call will succeed.

Reviewed by: Gleb Shchepa <gleb.shchepa@oracle.com>

Change-Id: I84ebcf4ee0ba9ceb89e1255bbb665c37cc1916a1