Bug #30909064 	EXPLAIN NEGLECTS DB PART OF IDENTIFIER AFTER CREATE VIEW
Contributed by Kaiwang CHen

The EXPLAIN output for a DML statement contains the table identifier
in the 'SHOW WARNINGS' output. This table identifier should normally
contain the database name. However, for certain commands such as
CREATE VIEW, the database name should be omitted, which is enforced
by setting the 'alias_name_used' flag to true in the cached table
object. If the cached table is reused after a CREATE VIEW, the
'alias_name_used' flag is not reset. This causes the database name to
be omitted from the EXPLAIN warnings for commands run after a CREATE
VIEW which access the same cached table as the view.

This is fixed by ensuring that the 'alias_name_used' flag is always
set to an appropriate value during table init. So the previous value
of this flag is overwritten to a value which is correct for the
current usage.

Change-Id: If7c861f28a5d9f97c57704a148dd475f39d0e7b6