WL#14630 Make UNIX_TIMESTAMP, FROM_UNIXTIME work with 64 bits time when available
This patch makes functional changes to platforms that have 64 bits
time for the system functions:

   FROM_UNIXTIME
   UNIX_TIMESTAMP
   CONVERT_TZ

will now allow values up UTC 3001-01-18 23:59:59, making use of the
wider range of the 64 bits time representation. This particular point
in time represents the maximum common point for Solaris, Linux, MaxOS
and Windows, the latter OS not allowing larger values.

SET TIMESTAMP will still be limited to 32 bits maximum value, since
data dictionary cannot create tables with times after 2038 (it uses
the SQL type TIMESTAMP, which is 32 bits, and unchanged by this
patch).

If a too large value is given to FROM_UNIXTIME, the result will be
NULL, for both 64 bits time platforms and 32 bits time platforms.  The
maximum values allowed are 32536771199 and 2147483647, respectively.

If a value with year larger than 9999 is given to UNIX_TIMESTAMP the
diagnostic [warning with SELECT) / error message (INSERT with strict
mode)] will be, as before, e.g.

    Incorrect datetime value: '10000-01-01 00:00.00'

The value returned will be 0.00000, with type of DECIMAL(17,6), as
before.  This applies to both 64 bits and 32 bits time platforms.

If a value larger than UTC 3001-01-18 23:59:59 is given to
UNIX_TIMESTAMP value will be 0 with no diagnostic and the type will be
BIGINT, for both 32 and 64 bits time platforms.

If a value larger than UTC 2038-01-19 03:14:07 is given to
UNIX_TIMESTAMP with 32 bits time, the value will be 0, with no
diagnostic, and the type will be BIGINT, again as before.

Notice the different type in the two cases: DECIMAL(17,6) vs
BIGINT.

Other user observable change:

CONVERT_TZ on 64 bit platforms will now handle times up to UTC
3001-01-18 23:59:59.

Internally, my_time_t will now always be 64 bits, with limits defined
by the new constants:

MYTIME_MIN_VALUE   0
MYTIME_MAX_VALUE   (64 bits platform: a value corresponding to
                                      3001-01-18 23:59:59
                    32 bits platform: the old value of TIMESTAMP_MAX_VALUE,
                                      i.e. max of signed int int32.

New constants that describe limits for the legacy 32 bits SQL type
TIMESTAMP are described by

    TYPE_TIMESTAMP_MIN_VALUE
    TYPE_TIMESTAMP_MAX_VALUE

The check for overflow is now performed in convert_TIME_to_timestamp.

We use our own struct timeval, since on 64 bits windows, the tv_sec
member is only 32 bits (typed long, which on Windows compiles to 32
bits int), except when interfacing with the host OS to read/set system
time.

The WL also closes Bug#11745263.

Change-Id: Ic8cef39dafe15896a8a82b57973374024098d550