Bug #32997832 REPLICATION IS BREAKING WITH  ...  ERROR 261 'DML COUNT IN TRANSACTI$
Changes in 8.0 result in a *permanent* error (261) being returned when a
transaction's operations do not fit in TC.

Previously this would normally have returned temporary error (233).
In the replication context, the temporary error will be retried, the
permanent error will not.

The permanent error is suitable when the *transaction itself* is too
large to *ever fit* in TC, without reconfiguration.

However in the case where the transaction cannot fit due to some other
transaction(s) consuming resources, the error should be temporary, as
it may fit later, or in some other TC instance.

The previously returned temporary error (233) now has a slightly different
meaning :
  In the past : Hit MaxNoOfConcurrentOperations
  From 8.0    : Insufficient pooled memory to allocate another operation

Rather than conflate these meanings (dynamic allocation + configured limit),
this patch adds a new Temporary error (234) which is returned when we
hit the configured limit.

Summary of new state :

 233 : Temporary : TC dynamic allocation failure - increase SharedGlobalMemory
 234 : Temporary : TC limit hit - increase MaxNoOfConcurrentOperations
 261 : Permanent : This transaction is too large wrt DML ops

testNdbApi -n MaxRows is modified to understand the new temporary error.

Approved-by : Mauritz Sundell <mauritz.sundell@oracle.com>