Bug #33104337 NDB_OPT.STATISTICS0/1 FAILS IN PB2
An 'attribute not found' error is returned on a pushed join in
NDB. The error is hit while looking up a column to add a linked
value. The lookup fails because the query uses the wrong case
on the column (colx vs. COLX).

This is fixed by using the original field name for lookup.

Change-Id: I8c760a9cbeaea6959f176f0a2ce8faa29dead3ab