Bug#33069443: MTR TESTS FAILING ON PB2 DAILY-TRUNK SINCE ROLLUP WORKLOG
Some of the mtr tests in rapid suite (group_min_max_innodb rollup_varlen
cp_rollup_varlen) are failing because Item_sum_count() is marked nullable
when it should not be . This happens because all functions
having a rollup expression as argument are marked nullable. However,
for rollup_sum_switchers, we do not need to make it nullable. The idea
of making a function nullable is because a rollup row generates
null for that expression. However for an aggregation function, rollup
row generates super aggregate value. So, there is no point in making
it nullable once the constructor (for Item_rollup_sum_switcher) sets it
as not nullable.

We now skip marking Item_rollup_sum_switcher items as nullable even
if the argument to these functions have a rollup expression.
They derive their nullability from the original aggregation function
for which sum switcher is created.

No test case on mysql side, as the nullability is not considered during
evaluation of the Item_rollup_sum_switcher. However, secondary engine seems
to depend on the information. Hence those tests pass with the change.

Change-Id: Id9a6ac60fdeb353c810292bf79c4a7041a680f6b