BUG#33025231: UNRESPONSIVE SERVER: SHELL HANGS FOR 2H IN ADDINSTANCE() (GR+CLONE)
ACL statements, like `CREATE USER` do acquire a exclusive lock on
`MDL_key::ACL_CACHE` from the beginning of the statement until it is
committed.  On Group Replication transactions are sent to the group
during the commit stage so that we have a global order, from which the
transaction GTID is inferred. The transactions are intercept on the
primary and will wait until the Group Communication Service (GCS)
acknowledges that the transactions was delivered to all members.  This
extends the time that the `MDL_key::ACL_CACHE` exclusive lock is
acquired, which means that during the start of the ACL statement, the
time GCS takes to deliver to all member and the ACL statement commits
locally, no other operations that do require any kind of lock on
`MDL_key::ACL_CACHE` will execute.

The solution for this scenario is to rewrite the steps that do read
the GTID_EXECUTED and GTID_PURGED to do not rely on SQL API, thence
not requiring a `MDL_key::ACL_CACHE` lock acquire. We rewrite the SQL
API using server services to access to system variables.

When XCOM uses MySQL stack for communication it will uses retries until
lock is released.

RB: 26671