Bug#33139598	REGRESSION: CRASH IN GET_BEST_GROUP_MIN_MAX
When a SEL_ARG is set to be a "MAYBE_KEY", even if the root of the range
tree is available, the call to "first()" would return nullptr. There is no
check for this before we access first()->next resulting in server exit.

We now check for first() before we access first()->next

Change-Id: Ie8419baca01df8c46ac0aff5b1cfc2d91acb162a