BUG#33082509: SHOW STATEMENTS SHOULD BE BLOCKED ON GR PRIMARY BEING RECOVERED
When Group Replication is configured with
group_replication_consistency = BEFORE_ON_PRIMARY_FAILOVER, if a
primary failover happens, client connections are hold until the new
primary has the same state of the previous primary. Otherwise, a
given connection that is transparently redirect to the new primary,
may not see their last updates or see a different time line.
Some statements are exempt from this hold because they are either
monitoring or administration statements.
SHOW statements are exempt, though some SHOW statements outcome do
depend on data, which means that those statements should be hold.

To solve the above issue, we handle SHOW statements in two ways.
SHOW statements that do depend on data are hold, being this the
default handling of SHOW statements.

The SHOW statements that only depend on status or configuration are
exempt from the hold, which are:
  SHOW VARIABLES;
  SHOW PROCESSLIST;
  SHOW STATUS;
  SHOW ENGINE INNODB LOGS;
  SHOW ENGINE INNODB STATUS;
  SHOW ENGINE INNODB MUTEX;
  SHOW MASTER STATUS;
  SHOW REPLICA STATUS;
  SHOW CHARACTER SET;
  SHOW COLLATION;
  SHOW BINARY LOGS;
  SHOW OPEN TABLES;
  SHOW REPLICAS;
  SHOW BINLOG EVENTS;
  SHOW WARNINGS;
  SHOW ERRORS;
  SHOW ENGINES;
  SHOW PRIVILEGES;
  SHOW PROCEDURE STATUS;
  SHOW FUNCTION STATUS;
  SHOW PLUGINS;
  SHOW EVENTS;
  SHOW PROFILE;
  SHOW PROFILES;
  SHOW RELAYLOG EVENTS;

RB: 26764