Bug#33104036: ASSERTION FAILURE IN FIND_SUBQUERY_IN_SELECT_LIST AT               ../SQL/ITEM_SUBSELECT.CC
Problem:
For a subquery which has a rollup wrapper like this :
Item_rollup_group_item(subquery), we first need to unwrap
the item to get the subquery. This subquery will then be
compared with the visible fields in the query to be
replaced later by the field from the derived table.

The reason for the assert is that, while comparing the
visible fields with the subquery, if its a match we
check if base_ref_item is matching with the item. It is
not in this case leading to the assert.

The reason being that after we unwrap the rollup item,
we are actually manupulating the original field list
itself instead of what was intended to be a manipulation
of the local object. For the query in question, the order
by field is first wrapped in a rollup group wrapper
(in resolve_rollup()), then it gets a relook in
find_order_in_list(). In find_order_in_list(), we add the order
by field as the hidden item and the field is marked as hidden.
Note that since its wrapped in rollup group item, it is the
rollup group item that gets the hidden flag and not the item
inside it. So when this is unwrapped to be compared and replaced
during transformation, we have a visible field. So this makes
the visible fields list longer and there by asserting since the
indexing w.r.t base_ref_items and the visible fields list
does not match anymore.

Solution is to unwrap the rollup group item everywhere (before,
after and in the call to replace_subquery_in_expr()) for
comparisons and replacements. We take care to retain the
rollup wrappers while the inner item gets replaced.

Change-Id: I496676560768d46cf8b5998bcde2bd0662035514