Bug#33079592: ASSERTION `SELECT->BASE_REF_ITEMS[ITEM_IDX]               == ITEM' FAILED Bug#32863279: STACK-OVERFLOW IN ITEM_FUNC_XOR WITH               SUBQUERY_TO_DERIVED=ON
The bug is about handling hidden items added as part of
order by or windowing functions when a transformation for
implicity grouped query is written. Here is a simple query:

SELECT SUM(i), ROW_NUMBER() OVER (), (SELECT MIN(i) FROM t1)
 FROM t1 ORDER BY i;

"i" in the order by is added as a hidden item to the query
block. The query is implicitly grouped, so there is no use
of order by. So, the order by can be removed. But we have
a window function. If the window function would have added
any hidden items (part of partition by or order by)
it would have done so after the order by items in the
ref_item_array. This makes it difficult to remove the order
by elements from the ref array. So we only remove the order
by list. But keep the field in the fields list.
During transformations (transform_grouped_to_derived()), field
"i" is added to the derived table. This becomes visible in
the derived table. We replace the field with the derived
table field which is now visible in the main query block.
This leads to the assert since the indexing in the
base_ref_item array and the visible fields list does not
match anymore when the subquery is getting replaced with
the derived table field in the same query block.

Solution:
Note the "hidden" flag for all expressions in the transforming
query block. When replacing expressions with derived table fields,
if the expression was marked hidden, we make a copy
of the derived table field to replace in this expression.
A copy is needed because we might have a situation where the
same derived table field could be used to replace another
expression which is marked visible in the query block.
For ex:
SELECT SUM(i), ROW_NUMBER(i) OVER (), i, (SELECT MIN(i) FROM t1)
FROM t1 ORDER BY i;
In this case, "i" is part of the derived table created.
However, "i" in the transforming query block has two occurances.
One added as hidden item during split_sum_func() for the aggregation
function and one "visible" in the query. For this case, we cannot
use the same instance of the derived table field to replace both
the "visible" field and the "hidden" field.
As a result, we use the copy for hidden expressions and mark the
expression as hidden.

Also, note that, without "subquery_to_derived" transformation,
first query passes. However it should throw error (postgres does)
which it does when the transformation is on. The full group
by checks happen after the order by is removed. Hence there are no
errors when the switch is off.

Bug#32863279: STACK-OVERFLOW IN ITEM_FUNC_XOR WITH
             SUBQUERY_TO_DERIVED=ON

Problem is similar to Bug#33079592, except that it is w.r.t
view refs. Note that we only need to copy the hidden flag
for a view ref as each of the view ref already gets a different
Item_field in replace_item_view_ref.

We also update base_ref_items when an expression(fields/view_refs)
in the transforming query block gets replaced with derived table field.

Change-Id: Ib67b132d153654325c9167cee2b2c99b5ea3c87a