Bug#33170381 - Port binding related failures in rest_api_enable test
Issue
=====
Some tests use the default https port for the REST API when starting a Router.
It causes those tests to fail if the port is used already.
Fix
===
Use a custom https ports. For testing the default confiiguration skip the
test if the default https port is not available.

RB: 26777
Reviewed by: Jan Kneschke <jan.kneschke@oracle.com>