WL#9149: MySQL GCS: Single Writer configuration
This worklog aligns XCom with the most used Group Replication (GR)
operation mode: single primary. XCom will be able to operate
with a single leader driving consensus.

We expect users to benefit from improved performance when they deploy
GR in single primary mode because GR will instruct XCom to operate
with a single "consensus leader" co-located with GR's primary.
We expect this improvement to be more apparent when some of
GR secondaries are unreachable.

This new behavior must be optional but in an opt-out fashion, since it
will be enabled by default.

RB: 24623