Bug#33048169: DISABLE REPLICA_PRESERVE_COMMIT_ORDER WHEN THE NUMBER OF WORKER = 1
The mechanism for preserving the commit order in a replica is
necessary when more than one thread are trying to commit transactions
concurrently.
If only one worker is reading and applying transactions, then this
mechanism is not needed and should not be instantiated.

This patch thus makes the replica not use this order insurance
mechanism if the server was configure to only use one MTA worker.

This patch also fixes some test dependencies on exact numbers of
workers that were found while testing this.

ReviewBoard: 26772