Bug #33149402: SELECT'S GROUPING() NOT TAKEN IN ACCOUNT IN HAVING WHEN JOIN USED
When setting up an Item_ref to a SELECT alias, we copy its cached
properties (including whether it is part of a ROLLUP expression or not).
However, these may not yet be correctly computed, so we need to do that
first, or we may get stuck with a wrong value. This can cause certain
expressions to be materialized in an intermediate step when they should not
(because they contain ROLLUP expressions that are not ready for
computation -- but we have the wrong cached values and don't know that).
Fix by forcing them to be recomputed when we designate an item as
a rollup item.

Change-Id: I5a660ab90cc607ec40b30805042e450a7dac251b