BUG#33082678 SOME MAJORITY LOSS TESTS FAIL SPORADICALLY ON LINUX
Description:
-----------
Some majority loss tests fail sporadically on Linux after WL#9149 code
is pushed

- group_replication.gr_majority_loss_restart_mysql
- group_replication.gr_acf_receiver_network_partition
- group_replication.gr_acf_sender_no_majority

Analysis and Suggested Fix:
----------------------------
From an initial analysis, the forced seems to take too long and it is deemed
as Duplicated View. On the correct cases, the view is considered duplicated,
but another view (non-duplicated) is delivered 30 seconds after the first
event.

This happens because we changed the way in which we force the
event horizon slots. Before this WL#9149, we forced the pipeline
with a 3p PAXOS until we get the next view installed, and then we
skip_op'ed after that. Now, to ensure that we have a crash safe approach,
we 3p PAXOS before and after the forced configuration is installed.

This causes a natural delay because 3p is slower than our optimized
implementation, thus causing the tests to timeout.

The fix was to increase the "Force Members" view timeout to 120s.

RB: 26812